Added fix to be able to add import to  campaign table  crowdfunding_db_schema.sql

+++ crowdfunding_db_schema.sql
@@ -3,23 +3,21 @@
 
 
 CREATE TABLE "campaign" (
-    "cf_id" INT   NOT NULL,
-    "contact_id" INT   NOT NULL,
+    "cf_id" INTEGER   NOT NULL,
+    "contact_id" INTEGER   NOT NULL,
     "company_name" VARCHAR(50)   NOT NULL,
     "description" VARCHAR(250)   NOT NULL,
-    "goal" NUMERIC(10,2)   NOT NULL,
-    "pledged" NUMERIC(10,2)   NOT NULL,
+    "goal" FLOAT  NOT NULL,
+    "pledged" FLOAT  NOT NULL,
     "outcome" VARCHAR(30)   NOT NULL,
-    "backers_count" INT   NOT NULL,
+    "backers_count" INTEGER   NOT NULL,
     "country" VARCHAR(100)   NOT NULL,
     "currency" VARCHAR(10)   NOT NULL,
     "launched_date" DATE   NOT NULL,
     "end_date" DATE   NOT NULL,
     "category_id" VARCHAR(10)   NOT NULL,
     "subcategory_id" VARCHAR(20)   NOT NULL,
-    CONSTRAINT "pk_campaign" PRIMARY KEY (
-        "cf_id"
-     )
+    CONSTRAINT "pk_campaign" PRIMARY KEY ("cf_id")
 );
 
 CREATE TABLE "category" (
